Remove Elasticsearch6SearchBackend and make Elasticsearch7SearchBackend the base implementation
